requirements.txt
Flask==0.12
Flask-RESTful==0.3.6
flask_sqlalchemy==2.1
flask_marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.2
psycopg2==2.7.3.2
FlaskAPI==1.0
flask_migrate==2.2.1